+++ src/actions/helpSomeoneActionCreators.js
@@ -34,3 +34,13 @@ export function setCurrentMember(member) {
 		member: member
 	}
 }
+
+
+
+export function getCharityFeaturedMembers(charity_id) {
+	const id = charity_id.toString
+	return {
+		type: 'GET_CHARITY_FEATURED_MEMBERS',
+		payload: axios.get('http://localhost:3000/charity_featured_members/'+ id)
+	}
+}
