+++ src/components/CharityPageHero/index.jsx
@@ -13,7 +13,7 @@ import CareButton from '../CareButton'
 
 class CharityPageHero extends React.Component {
 
- 
+
 
   render() {
     return (
@@ -21,7 +21,7 @@ class CharityPageHero extends React.Component {
       <ModalVideo channel='youtube' isOpen={this.props.videoVisible} videoId='xrCJa7iWnWY' />
       <div className="charity-banner-wrapper">
         <img className="charity-banner" src="images/streetworkbanner.jpg" />
-        <div className="linear-gradient"></div> 
+        <div className="linear-gradient"></div>
       </div>
       <img className="charity-logo" src="images/streetwork.jpg" />
       <h2>Streetwork</h2>
@@ -31,7 +31,7 @@ class CharityPageHero extends React.Component {
         <a href="https://www.facebook.com/StreetChange1/" target="_blank"><i className="fa fa-twitter"></i></a>
         <a href="https://www.facebook.com/StreetChange1/" target="_blank"><i className="fa fa-globe"></i></a>
       </div>
-    <button className="watch-a-video" onClick={this.props.toggleVideo}>Watch a Video<i class="fa fa-play-circle-o" aria-hidden="true"></i></button> 
+    <button className="watch-a-video" onClick={this.props.toggleVideo}>Watch a Video<i className="fa fa-play-circle-o" aria-hidden="true"></i></button> 
     <CareButton />
     </div>
 		)
