+++ src/reducers/charityPage.js
@@ -7,13 +7,6 @@ function charityPage(state = [], action) {
     case 'TOGGLE_VIDEO':
 	   	return {...state, videoVisible: !state.videoVisible}
 
-		case 'GET_CHARITY_FEATURED_MEMBERS_PENDING':
-	    return {...state, fetching: true}
-		case 'GET_CHARITY_FEATURED_MEMBERS_REJECTED':
-	    return {...state, fetching: true, error: action.payload}
-		case 'GET_CHARITY_FEATURED_MEMBERS_FULFILLED':
-	    return {...state, fetching: true, fetched: true, members: action.payload.data}
-
 		default:
 			return state
